// backend/src/api/auth/constants/errorMessages.ts
export enum ErrorMessages {
  USER_ALREADY_EXISTS = "A user with this email already exists",
  INVALID_EMAIL_OR_PASSWORD = "Invalid email or password",
  REFRESH_TOKEN_NOT_FOUND = "Refresh token not found",
  INVALID_REFRESH_TOKEN = "Invalid refresh token",
  SIGNUP_UNEXPECTED_ERROR = "An unexpected error occurred during signup",
  SIGNIN_UNEXPECTED_ERROR = "An unexpected error occurred during signin",
  REFRESH_UNEXPECTED_ERROR = "An unexpected error occurred during token refresh",
  LOGOUT_UNEXPECTED_ERROR = "An unexpected error occurred during logout",
}